moved merge settings into utilities
